adding extra control
git-svn-id: file:///home/svn/shoola-omero/trunk@7776 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/metadata/view/MetadataViewerComponent.java
@@ -506,7 +506,8 @@ class MetadataViewerComponent
 				b = false;
 			}	
 		}
-		deleteAnnotations(toDelete);
+		if (toDelete != null && toDelete.size() > 0)
+			deleteAnnotations(toDelete);
 		if (b) fireStateChange();
 	}
